# api/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Private REST API to for user interface application",
        "title": "storePredictor private REST API",
        "contact": {
            "name": "Ales Jandera",
            "url": "http://storepredictor.com",
            "email": "ales@storepredictor.com"
        },
        "license": {
            "name": "Commercial"
        },
        "version": "1.0"
    },
    "host": "localhost:9996",
    "basePath": "/",
    "paths": {
...
        "/account/order/{accountId}/{storeId}": {
            "get": {
                "description": "Endpoint to get account orders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Endpoint to get account orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            },
            "post": {
                "description": "Endpoint to store Order data from account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Endpoint to store Order data from account",
                "parameters": [
                    {
                        "description": "Order object to store in database",
                        "name": "o.Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
...
        "/order": {
            "post": {
                "description": "Endpoint to store Order data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Endpoint to store Order data",
                "parameters": [
                    {
                        "description": "Order object to store in database",
                        "name": "o.Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
        "/orders/product/{storeId}/{productCode}": {
            "get": {
                "description": "Endpoint to get Orders for each product by code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Endpoint to get Orders for each product by code",
                "parameters": [
                    {
                        "description": "Order object to store in database",
                        "name": "o.Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
        "/orders/{storeId}/{limit}/{offset}": {
            "get": {
                "description": "Endpoint to get orders for each Client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Endpoint to get orders for each Client",
                "parameters": [
                    {
                        "description": "Order object to store in database",
                        "name": "o.Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
...
        "/prediction/manual/{storeId}/{date}": {
            "post": {
                "description": "Endpoint to start manualy prediction for set day",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Endpoint to start manualy prediction for set day",
                "parameters": [
                    {
                        "description": "Order object to store in database",
                        "name": "o.Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
        "/prediction/onetime/{orderId}": {
            "post": {
                "description": "Endpoint to store Order for one time prediction data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Endpoint to store Order for one time prediction data",
                "parameters": [
                    {
                        "description": "Order object to store in database",
                        "name": "o.Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
...
        "/upgrade/{accountId}": {
            "post": {
                "description": "Endpoint to store Order data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Endpoint to store Order data",
                "parameters": [
                    {
                        "description": "Order object to store in database",
                        "name": "o.Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Order": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "companyNumber": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paidTo": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "planRefer": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "vatNumber": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
...
        "utils.Res": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}